Refactor load_config
